Revert "refactor directories"
This reverts commit 7018e0652841b1a645d708837e5d3bbb53e452a0.
