+++ src/models/collector.ts
@@ -1,12 +1,12 @@
 'use strict'
 import { Model, DataTypes, Sequelize, Op } from 'sequelize'
-import { Citadel } from '../citadel'
-import { State } from '../state'
-import { LevelFactory } from './level-factory'
-import { RepairFactory } from './repair-factory'
-import { WeatherPlayer } from '../weather-player'
+import { Citadel } from './citadel'
+import { State } from './state'
+import { LevelCollector } from './level-collector'
+import { RepairCollector } from './repair-collector'
+import { WeatherPlayer } from './weather-player'
 
-export class Factory extends Model {
+export class Collector extends Model {
   public idCitadel!: number
   public idState!: number
   public idNextState!: number | null
@@ -14,41 +14,41 @@ export class Factory extends Model {
   public finishTime!: Date | null
   public health!: number
   public state?: State
-  public levelFactory?: LevelFactory
+  public levelCollector?: LevelCollector
+  public repairCollector?: RepairCollector
   public citadel?: Citadel
-  public repairFactory?: RepairFactory
 
   static associate(models: any) {
-    Factory.belongsTo(models.Citadel, {
+    Collector.belongsTo(models.Citadel, {
       foreignKey: 'idCitadel',
       targetKey: 'id',
       as: 'citadel',
     })
-    Factory.hasOne(models.State, {
+    Collector.hasOne(models.State, {
       foreignKey: 'id',
       sourceKey: 'idState',
       as: 'state',
     })
-    Factory.hasOne(models.LevelFactory, {
+    Collector.hasOne(models.LevelCollector, {
       foreignKey: 'level',
       sourceKey: 'level',
-      as: 'levelFactory',
+      as: 'levelCollector',
     })
-    Factory.hasOne(models.RepairFactory, {
+    Collector.hasOne(models.RepairCollector, {
       foreignKey: 'idStateFrom',
       sourceKey: 'idState',
-      as: 'repairFactory',
+      as: 'repairCollector',
     })
   }
 
-  async updateFactory(db: any, date: Date) {
+  async updateCollector(db: any, date: Date) {
     if (!this.finishTime) {
       return
     }
     if (this.finishTime < date) {
-      if (this.state!.isManufacturing()) {
-        const yieldFactory = this.levelFactory!.yield
-        await this.citadel!.addMaterials(yieldFactory)
+      if (this.state!.isCollecting()) {
+        const yieldCollector = this.levelCollector!.yield
+        await this.citadel!.addResources(yieldCollector)
       } else if (this.state!.isUpgrading()) {
         this.level += 1
       }
@@ -60,96 +60,89 @@ export class Factory extends Model {
     }
   }
 
-  async manufacture(db: any) {
+  async collect(db: any) {
     if (this.finishTime) {
-      throw new Error('Factory busy')
+      throw new Error('Collector busy')
     }
-    if (this.citadel!.resources < this.levelFactory!.resources) {
-      throw new Error('Not enough resources')
-    }
-
-    this.citadel!.resources -= this.levelFactory!.resources
-    await this.citadel!.save()
-
-    const timeToManufactor = this.levelFactory!.timeToManufactor
-    const [hours, minutes, seconds] = timeToManufactor.split(':').map(Number)
+    const timeToCollect = this.levelCollector!.timeToCollect
+    const [hours, minutes, seconds] = timeToCollect.split(':').map(Number)
     const milliseconds = (hours * 3600 + minutes * 60 + seconds) * 1000
     this.finishTime = new Date(Date.now() + milliseconds)
     this.idNextState = this.idState
-    const manufacturingState = await db.State.getManufacturingState()
-    this.idState = manufacturingState.id
+    const collectingState = await db.State.getCollectingState()
+    this.idState = collectingState.id
 
     await this.save()
   }
 
   async repair(db: any) {
     if (!this.state!.canRepair()) {
-      throw new Error('Factory cannot be repaired')
+      throw new Error('Collector cannot be repaired')
     }
-    if (this.citadel!.resources < this.repairFactory!.resources) {
+    if (this.citadel!.resources < this.repairCollector!.resources) {
       throw new Error('Not enough resources to repair')
     }
-    if (this.citadel!.materials < this.repairFactory!.materials) {
+    if (this.citadel!.materials < this.repairCollector!.materials) {
       throw new Error('Not enough materials to repair')
     }
 
-    this.citadel!.resources -= this.repairFactory!.resources
-    this.citadel!.materials -= this.repairFactory!.materials
+    this.citadel!.resources -= this.repairCollector!.resources
+    this.citadel!.materials -= this.repairCollector!.materials
     await this.citadel!.save()
 
-    const timeToRepair = this.repairFactory!.timeToRepair
+    const timeToRepair = this.repairCollector!.timeToRepair
     const [hours, minutes, seconds] = timeToRepair.split(':').map(Number)
     const milliseconds = (hours * 3600 + minutes * 60 + seconds) * 1000
     this.finishTime = new Date(Date.now() + milliseconds)
     const repairState = await db.State.getRepairingState()
     this.idState = repairState.id
-    this.idNextState = this.repairFactory!.idStateTo
+    this.idNextState = this.repairCollector!.idStateTo
 
     await this.save()
   }
 
   async reinforce(db: any) {
     if (!this.state!.canReinforce()) {
-      throw new Error('Factory cannot be reinforced')
+      throw new Error('Collector cannot be reinforced')
     }
-    if (this.citadel!.resources < this.repairFactory!.resources) {
+    if (this.citadel!.resources < this.repairCollector!.resources) {
       throw new Error('Not enough resources to reinforce')
     }
-    if (this.citadel!.materials < this.repairFactory!.materials) {
+    if (this.citadel!.materials < this.repairCollector!.materials) {
       throw new Error('Not enough materials to reinforce')
     }
 
-    this.citadel!.resources -= this.repairFactory!.resources
-    this.citadel!.materials -= this.repairFactory!.materials
+    this.citadel!.resources -= this.repairCollector!.resources
+    this.citadel!.materials -= this.repairCollector!.materials
     await this.citadel!.save()
 
-    const timeToReinforce = this.repairFactory!.timeToRepair
+    const timeToReinforce = this.repairCollector!.timeToRepair
     const [hours, minutes, seconds] = timeToReinforce.split(':').map(Number)
     const milliseconds = (hours * 3600 + minutes * 60 + seconds) * 1000
     this.finishTime = new Date(Date.now() + milliseconds)
     const reinforceState = await db.State.getRepairingState()
     this.idState = reinforceState.id
-    this.idNextState = this.repairFactory!.idStateTo
+    this.idNextState = this.repairCollector!.idStateTo
 
     await this.save()
   }
 
   async upgrade(db: any) {
     if (!this.state!.canUpgrade()) {
-      throw new Error('Factory cannot be upgraded')
+      throw new Error('Collector cannot be upgraded')
     }
-    if (this.citadel!.resources < this.levelFactory!.upgradeResources) {
+    if (this.citadel!.resources < this.levelCollector!.upgradeResources) {
       throw new Error('Not enough resources to upgrade')
     }
-    if (this.citadel!.materials < this.levelFactory!.upgradeMaterials) {
+    if (this.citadel!.materials < this.levelCollector!.upgradeMaterials) {
       throw new Error('Not enough materials to upgrade')
     }
 
-    this.citadel!.resources -= this.levelFactory!.upgradeResources
-    this.citadel!.materials -= this.levelFactory!.upgradeMaterials
+    this.citadel!.resources -= this.levelCollector!.upgradeResources
+    this.citadel!.materials -= this.levelCollector!.upgradeMaterials
     await this.citadel!.save()
 
-    const timeToUpgrade = this.levelFactory!.timeToUpgrade
+    const timeToUpgrade = this.levelCollector!.timeToUpgrade
     const [hours, minutes, seconds] = timeToUpgrade.split(':').map(Number)
     const milliseconds = (hours * 3600 + minutes * 60 + seconds) * 1000
     this.finishTime = new Date(Date.now() + milliseconds)
@@ -198,7 +191,7 @@ export class Factory extends Model {
 }
 
 export default (sequelize: Sequelize) => {
-  Factory.init(
+  Collector.init(
     {
       idCitadel: {
         type: DataTypes.INTEGER,
@@ -226,10 +219,10 @@ export default (sequelize: Sequelize) => {
     },
     {
       sequelize,
-      modelName: 'Factory',
-      tableName: 'Factory',
+      modelName: 'Collector',
+      tableName: 'Collector',
     }
   )
 
-  return Factory
+  return Collector
 }
