+++ src/models/factory.ts
@@ -1,12 +1,12 @@
 'use strict'
 import { Model, DataTypes, Sequelize, Op } from 'sequelize'
-import { Citadel } from '../citadel'
-import { State } from '../state'
-import { LevelCollector } from './level-collector'
-import { RepairCollector } from './repair-collector'
-import { WeatherPlayer } from '../weather-player'
+import { Citadel } from './citadel'
+import { State } from './state'
+import { LevelFactory } from './level-factory'
+import { RepairFactory } from './repair-factory'
+import { WeatherPlayer } from './weather-player'
 
-export class Collector extends Model {
+export class Factory extends Model {
   public idCitadel!: number
   public idState!: number
   public idNextState!: number | null
@@ -14,41 +14,41 @@ export class Collector extends Model {
   public finishTime!: Date | null
   public health!: number
   public state?: State
-  public levelCollector?: LevelCollector
-  public repairCollector?: RepairCollector
+  public levelFactory?: LevelFactory
   public citadel?: Citadel
+  public repairFactory?: RepairFactory
 
   static associate(models: any) {
-    Collector.belongsTo(models.Citadel, {
+    Factory.belongsTo(models.Citadel, {
       foreignKey: 'idCitadel',
       targetKey: 'id',
       as: 'citadel',
     })
-    Collector.hasOne(models.State, {
+    Factory.hasOne(models.State, {
       foreignKey: 'id',
       sourceKey: 'idState',
       as: 'state',
     })
-    Collector.hasOne(models.LevelCollector, {
+    Factory.hasOne(models.LevelFactory, {
       foreignKey: 'level',
       sourceKey: 'level',
-      as: 'levelCollector',
+      as: 'levelFactory',
     })
-    Collector.hasOne(models.RepairCollector, {
+    Factory.hasOne(models.RepairFactory, {
       foreignKey: 'idStateFrom',
       sourceKey: 'idState',
-      as: 'repairCollector',
+      as: 'repairFactory',
     })
   }
 
-  async updateCollector(db: any, date: Date) {
+  async updateFactory(db: any, date: Date) {
     if (!this.finishTime) {
       return
     }
     if (this.finishTime < date) {
-      if (this.state!.isCollecting()) {
-        const yieldCollector = this.levelCollector!.yield
-        await this.citadel!.addResources(yieldCollector)
+      if (this.state!.isManufacturing()) {
+        const yieldFactory = this.levelFactory!.yield
+        await this.citadel!.addMaterials(yieldFactory)
       } else if (this.state!.isUpgrading()) {
         this.level += 1
       }
@@ -60,89 +60,96 @@ export class Collector extends Model {
     }
   }
 
-  async collect(db: any) {
+  async manufacture(db: any) {
     if (this.finishTime) {
-      throw new Error('Collector busy')
+      throw new Error('Factory busy')
     }
-    const timeToCollect = this.levelCollector!.timeToCollect
-    const [hours, minutes, seconds] = timeToCollect.split(':').map(Number)
+    if (this.citadel!.resources < this.levelFactory!.resources) {
+      throw new Error('Not enough resources')
+    }
+
+    this.citadel!.resources -= this.levelFactory!.resources
+    await this.citadel!.save()
+
+    const timeToManufactor = this.levelFactory!.timeToManufactor
+    const [hours, minutes, seconds] = timeToManufactor.split(':').map(Number)
     const milliseconds = (hours * 3600 + minutes * 60 + seconds) * 1000
     this.finishTime = new Date(Date.now() + milliseconds)
     this.idNextState = this.idState
-    const collectingState = await db.State.getCollectingState()
-    this.idState = collectingState.id
+    const manufacturingState = await db.State.getManufacturingState()
+    this.idState = manufacturingState.id
 
     await this.save()
   }
 
   async repair(db: any) {
     if (!this.state!.canRepair()) {
-      throw new Error('Collector cannot be repaired')
+      throw new Error('Factory cannot be repaired')
     }
-    if (this.citadel!.resources < this.repairCollector!.resources) {
+    if (this.citadel!.resources < this.repairFactory!.resources) {
       throw new Error('Not enough resources to repair')
     }
-    if (this.citadel!.materials < this.repairCollector!.materials) {
+    if (this.citadel!.materials < this.repairFactory!.materials) {
       throw new Error('Not enough materials to repair')
     }
 
-    this.citadel!.resources -= this.repairCollector!.resources
-    this.citadel!.materials -= this.repairCollector!.materials
+    this.citadel!.resources -= this.repairFactory!.resources
+    this.citadel!.materials -= this.repairFactory!.materials
     await this.citadel!.save()
 
-    const timeToRepair = this.repairCollector!.timeToRepair
+    const timeToRepair = this.repairFactory!.timeToRepair
     const [hours, minutes, seconds] = timeToRepair.split(':').map(Number)
     const milliseconds = (hours * 3600 + minutes * 60 + seconds) * 1000
     this.finishTime = new Date(Date.now() + milliseconds)
     const repairState = await db.State.getRepairingState()
     this.idState = repairState.id
-    this.idNextState = this.repairCollector!.idStateTo
+    this.idNextState = this.repairFactory!.idStateTo
 
     await this.save()
   }
 
   async reinforce(db: any) {
     if (!this.state!.canReinforce()) {
-      throw new Error('Collector cannot be reinforced')
+      throw new Error('Factory cannot be reinforced')
     }
-    if (this.citadel!.resources < this.repairCollector!.resources) {
+    if (this.citadel!.resources < this.repairFactory!.resources) {
       throw new Error('Not enough resources to reinforce')
     }
-    if (this.citadel!.materials < this.repairCollector!.materials) {
+    if (this.citadel!.materials < this.repairFactory!.materials) {
       throw new Error('Not enough materials to reinforce')
     }
 
-    this.citadel!.resources -= this.repairCollector!.resources
-    this.citadel!.materials -= this.repairCollector!.materials
+    this.citadel!.resources -= this.repairFactory!.resources
+    this.citadel!.materials -= this.repairFactory!.materials
     await this.citadel!.save()
 
-    const timeToReinforce = this.repairCollector!.timeToRepair
+    const timeToReinforce = this.repairFactory!.timeToRepair
     const [hours, minutes, seconds] = timeToReinforce.split(':').map(Number)
     const milliseconds = (hours * 3600 + minutes * 60 + seconds) * 1000
     this.finishTime = new Date(Date.now() + milliseconds)
     const reinforceState = await db.State.getRepairingState()
     this.idState = reinforceState.id
-    this.idNextState = this.repairCollector!.idStateTo
+    this.idNextState = this.repairFactory!.idStateTo
 
     await this.save()
   }
 
   async upgrade(db: any) {
     if (!this.state!.canUpgrade()) {
-      throw new Error('Collector cannot be upgraded')
+      throw new Error('Factory cannot be upgraded')
     }
-    if (this.citadel!.resources < this.levelCollector!.upgradeResources) {
+    if (this.citadel!.resources < this.levelFactory!.upgradeResources) {
       throw new Error('Not enough resources to upgrade')
     }
-    if (this.citadel!.materials < this.levelCollector!.upgradeMaterials) {
+    if (this.citadel!.materials < this.levelFactory!.upgradeMaterials) {
       throw new Error('Not enough materials to upgrade')
     }
 
-    this.citadel!.resources -= this.levelCollector!.upgradeResources
-    this.citadel!.materials -= this.levelCollector!.upgradeMaterials
+    this.citadel!.resources -= this.levelFactory!.upgradeResources
+    this.citadel!.materials -= this.levelFactory!.upgradeMaterials
     await this.citadel!.save()
 
-    const timeToUpgrade = this.levelCollector!.timeToUpgrade
+    const timeToUpgrade = this.levelFactory!.timeToUpgrade
     const [hours, minutes, seconds] = timeToUpgrade.split(':').map(Number)
     const milliseconds = (hours * 3600 + minutes * 60 + seconds) * 1000
     this.finishTime = new Date(Date.now() + milliseconds)
@@ -191,7 +198,7 @@ export class Collector extends Model {
 }
 
 export default (sequelize: Sequelize) => {
-  Collector.init(
+  Factory.init(
     {
       idCitadel: {
         type: DataTypes.INTEGER,
@@ -219,10 +226,10 @@ export default (sequelize: Sequelize) => {
     },
     {
       sequelize,
-      modelName: 'Collector',
-      tableName: 'Collector',
+      modelName: 'Factory',
+      tableName: 'Factory',
     }
   )
 
-  return Collector
+  return Factory
 }
