+++ src/models/machine-gun-turret.ts
