+++ src/models/weather-forecast.ts
